Add triggers and next dialogue play

+++ Source/VoidSpace/DialogueLookupTable.h
@@ -31,5 +31,5 @@ public:
 	FString NextDialogue;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = Dialogue)
-	int NextDialogueDelay;
+	float NextDialogueDelay;
 };
